Minor perf improvements to frida-drcov.py
I started making some perf fixes before realizing some stuff about
frida.

So, the good news: these perf fixes do in fact speed up the client by
~10% or so.
The bad news: this tracer is still 2 orders of magnitude slower than
native execution. From memory, DR is approx 33% slower than native, and
pin is ~%50. So this is quite a bit worse. Intuitively, this makes sense
as all coverage data has to be IPC'd out to the `recv()`ing process.
There are ways this could be improved:
* Cache the bbs in frida address space and only IPC them out on client
detatch. However, if the client `exit()`'s or crashes, all coverage data
would be lost.
* Reduce allocations and copies inside the frida script. This is
complicated by javascripts lack of native u64's. I don't really know
javascript well enough to effectively optimize this.
* The one thats probably a good idea: Right now the javascript creates
the drcov objects and sends them to the python, which just is
responsible for uniquing them. If IPC is the big bottleneck, if we
uniqued them on the javascript side, we would reduce that and in theory
improve performance. Some initial tests of this were not successful, so
I'm going to have to play around with this more.

For what it's worth, frida offers two modes for tracing at the
granularity we care about: `compile` and `block`. `compile` will fire an
event the first time a bb is seen, whereas `block` will fire an event
every time a bb is seen. Switching to `compile` improves perf by an
order of magnitude, however any blocks that were hit before you attach
that you re-execute will not show up in your trace. It's pretty easy to
toggle between them, and if people think making this an option has
value, I'm not opposed. My thoughts were that in either case it's still
really slow in comparison, so we might as well be slow, accurate, and
easy to use.

So yeah, perf improvements, but they don't really matter \o/