Make coverage saving more robust
The most bug prone part of frida (in our usage) seems to be process
detatching. Previously the script would detatch and then persist all
coverage data once detatching had succeeded. This of course means that
if detatching fails or hangs, all collected coverage is lost. This is
Not Good.

This commit adds a SIGINT handler which persists the existing coverage
before calling `_exit(1)`. While this doesn't seem to be an especially
quick operation, it does save the coverage before exiting.

This commit does not address Windows usage, or processes exiting in ways
other than SIGINT.